+++ Bonobo.Git.Server/Controllers/RepositoryController.cs
@@ -431,6 +431,43 @@ namespace Bonobo.Git.Server.Controllers
                 var commits = browser.GetCommits(name, page, 10, out referenceName, out totalCount);
                 PopulateBranchesData(browser, referenceName);
                 ViewBag.TotalCount = totalCount;
+
+                var linksreg = repo.LinksUseGlobal ? UserConfiguration.Current.LinksRegex : repo.LinksRegex;
+                var linksurl = repo.LinksUseGlobal ? UserConfiguration.Current.LinksUrl : repo.LinksUrl;
+                foreach (var commit in commits)
+                {
+                    var links = new List<string>();
+                    if (!string.IsNullOrEmpty(linksreg))
+                    {
+                        try
+                        {
+                            var matches = Regex.Matches(commit.Message, linksreg);
+                            if (matches.Count > 0)
+                            {
+                                foreach (Match match in matches)
+                                {
+                                    IEnumerable<Group> groups = match.Groups.Cast<Group>();
+                                    var link = "";
+                                    try
+                                    {
+                                        var m = groups.Select(x => x.ToString()).ToArray();
+                                        link = string.Format(linksurl, m);
+                                    }
+                                    catch (FormatException e)
+                                    {
+                                        link = "An error occured while trying to format the link. Exception: " + e.Message;
+                                    }
+                                    links.Add(link);
+                                }
+                            }
+                        }
+                        catch (ArgumentException e)
+                        {
+                            links.Add("An error occured while trying to match the regualar expression. Error: " + e.Message);
+                        }
+                    }
+                    commit.Links = links;
+                }
                 return View(new RepositoryCommitsModel { Commits = commits, Name = repo.Name });
             }
         }
@@ -613,6 +650,9 @@ namespace Bonobo.Git.Server.Controllers
                 Status = GetRepositoryStatus(model),
                 AuditPushUser = model.AuditPushUser,
                 Logo = new RepositoryLogoDetailModel(model.Logo),
+                LinksUseGlobal = model.LinksUseGlobal,
+                LinksRegex = model.LinksRegex,
+                LinksUrl = model.LinksUrl,
             };
         }
 
@@ -644,7 +684,10 @@ namespace Bonobo.Git.Server.Controllers
                 AuditPushUser = model.AuditPushUser,
                 Logo = model.Logo != null ? model.Logo.BinaryData : null,
                 AllowAnonymousPush = model.AllowAnonymousPush,
-                RemoveLogo = model.Logo != null && model.Logo.RemoveLogo
+                RemoveLogo = model.Logo != null && model.Logo.RemoveLogo,
+                LinksUseGlobal = model.LinksUseGlobal,
+                LinksRegex = model.LinksRegex ?? "",
+                LinksUrl = model.LinksUrl ?? ""
             };
         }