+++ Bonobo.Git.Server/Data/Update/UpdateScriptRepository.cs
@@ -22,7 +22,8 @@ namespace Bonobo.Git.Server.Data.Update
                         new Sqlite.AddGroup(),
                         new Sqlite.AddRepositoryLogo(),
                         new Sqlite.AddGuidColumn(),
-                        new Sqlite.AddRepoPushColumn()
+                        new Sqlite.AddRepoPushColumn(),
+                        new Sqlite.AddRepoLinksColumn()
                     };
                 case "SqlConnection":
                     return new List<IUpdateScript>
@@ -34,7 +35,8 @@ namespace Bonobo.Git.Server.Data.Update
                         new SqlServer.AddGroup(),
                         new SqlServer.AddRepositoryLogo(),
                         new SqlServer.AddGuidColumn(),
-                        new SqlServer.AddRepoPushColumn()
+                        new SqlServer.AddRepoPushColumn(),
+                        new SqlServer.AddRepoLinksColumn()
                     };
                 default:
                     throw new NotImplementedException(string.Format("The provider '{0}' is not supported yet", sqlProvider));