+++ Bonobo.Git.Server/Models/SettingsModels.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Web;
 using System.ComponentModel.DataAnnotations;
 using Bonobo.Git.Server.App_GlobalResources;
+using Bonobo.Git.Server.Attributes;
 
 namespace Bonobo.Git.Server.Models
 {
@@ -43,6 +44,7 @@ namespace Bonobo.Git.Server.Models
         [Display(ResourceType = typeof(Resources), Name = "Settings_Global_LinksUrl")]
         public string LinksUrl { get; set; }
 
+        [IsValidRegex]
         [Display(ResourceType = typeof(Resources), Name = "Settings_Global_LinksRegex")]
         public string LinksRegex { get; set; }
     }